add validators
