+++ datamodel_code_generator/util.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, TypeVar
 
 import pydantic
 from packaging import version
 
-PYDANTIC_VERSION = version.parse(pydantic.VERSION)
+PYDANTIC_VERSION = version.parse(
+    pydantic.VERSION if isinstance(pydantic.VERSION, str) else str(pydantic.VERSION)
+)
 
 PYDANTIC_V2: bool = PYDANTIC_VERSION >= version.parse('2.0b3')
 
@@ -16,12 +18,13 @@ if TYPE_CHECKING:
 
     Protocol = object
     runtime_checkable: Callable[..., Any]
+    from pydantic import BaseModel
     from yaml import SafeLoader
 else:
     try:
         from typing import TYPE_CHECKING, Any, Callable, Protocol
     except ImportError:
-        from typing_extensions import Protocol  # noqa
+        from typing_extensions import Protocol, Literal  # noqa
     try:
         from typing import runtime_checkable
     except ImportError:
@@ -51,3 +54,39 @@ else:
 SafeLoader.yaml_constructors[
     'tag:yaml.org,2002:timestamp'
 ] = SafeLoader.yaml_constructors['tag:yaml.org,2002:str']
+
+
+Model = TypeVar('Model', bound=BaseModel)
+
+
+def model_validator(
+    mode: Literal['before', 'after'],
+) -> Callable[[Callable[[Model, Any], Any]], Callable[[Model, Any], Any]]:
+    def inner(method: Callable[[Model, Any], Any]) -> Callable[[Model, Any], Any]:
+        if PYDANTIC_V2:
+            from pydantic import model_validator as model_validator_v2
+
+            return model_validator_v2(mode=mode)(method)  # type: ignore
+        else:
+            from pydantic import root_validator
+
+            return root_validator(method)  # type: ignore
+
+    return inner
+
+
+def field_validator(
+    field_name: str,
+    mode: Literal['before', 'after'] = 'after',
+) -> Callable[[Callable[[Model, Any], Any]], Callable[[Model, Any], Any]]:
+    def inner(method: Callable[[Model, Any], Any]) -> Callable[[Model, Any], Any]:
+        if PYDANTIC_V2:
+            from pydantic import field_validator as field_validator_v2
+
+            return field_validator_v2(field_name, mode=mode)(method)  # type: ignore
+        else:
+            from pydantic import validator
+
+            return validator(field_name, pre=mode == 'before')(method)  # type: ignore
+
+    return inner
